refactor: update controller to add response entity all mappings

+++ src/main/java/dev/reso/practicing/controller/ClientController.java
@@ -1,9 +1,9 @@
 package dev.reso.practicing.controller;
 
 import dev.reso.practicing.dto.ClientDTO;
-import dev.reso.practicing.model.Client;
 import dev.reso.practicing.service.ClientService;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
@@ -23,24 +23,43 @@ public class ClientController {
     }
 
     @GetMapping("/{id}")
-    public ResponseEntity<ClientDTO> getById(@PathVariable Long id){
+    public ResponseEntity<?> getById(@PathVariable Long id){
         ClientDTO client = clientService.getById(id);
-        return ResponseEntity.ok().body(client);
+
+        if(client != null){
+            return ResponseEntity.ok(client);
+        }else{
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Not Found");
+        }
     }
 
     @PostMapping
     public ResponseEntity<ClientDTO>  insertClient(@RequestBody ClientDTO client){
        ClientDTO clientDTO =  clientService.insert(client);
-       return ResponseEntity.ok().body(clientDTO);
+       return ResponseEntity.status(HttpStatus.CREATED).body(clientDTO);
     }
 
     @DeleteMapping("/{id}")
-    public void deleteClient(@PathVariable Long id){
-        clientService.delete(id);
+    public ResponseEntity<String> deleteClient(@PathVariable Long id){
+
+        if(clientService.getById(id) != null){
+            clientService.delete(id);
+            return ResponseEntity.noContent().build();
+        }else{
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Ninja com id " + id + " nao encontrado!");
+        }
+
     }
 
     @PutMapping("/{id}")
-    public ClientDTO updateClient(@RequestBody ClientDTO client, @PathVariable Long id){
-        return clientService.update(client,id);
+    public ResponseEntity<?>  updateClient(@RequestBody ClientDTO client, @PathVariable Long id){
+        ClientDTO clientDTO =  clientService.update(client,id);
+        if(clientDTO != null){
+            return ResponseEntity.status(HttpStatus.CREATED).body(clientDTO);
+        }else{
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Not Found");
+        }
+
+
     }
 }
